10th

+++ graph/cycle.py
@@ -15,37 +15,31 @@ class Cycle:
         self.id = id
         self.id_text_guid = None
         self.cycle_mesh_guid = None
-        self.centroid = None
-        self.cycle = cycle
+        self.cycle_polyline = None
+        self.cycle_polyline_guid = None
+        self.centroid = None  # サイクルの構成点が5つ以上の場合は求めない
+        self.cycle = cycle  # ex. [4, 2, 0]
         self.composition_nodes = composition_nodes  # node instance
         self.is_on_GL = is_on_gl
 
     def generate_cycle_mesh(self, layer_name):
+
         # Generating cycle mesh
         cycle_mesh = Mesh()
 
-        print("composition_nodes num: {0}".format(len(self.composition_nodes)))
-
+        # メッシュの頂点情報を与える
         for node in self.composition_nodes:
             cycle_mesh.Vertices.Add(node.point)
 
+        # メッシュの面情報を与える
         if len(self.composition_nodes) == 3:
             cycle_mesh.Faces.AddFace(0, 1, 2)
         elif len(self.composition_nodes) == 4:
             cycle_mesh.Faces.AddFace(0, 1, 2, 3)
-
-        else:  # メッシュの構成点が5つ以上の場合
-            # TODO メッシュの最小分割面は三角形になるので、それを念頭においておく
-            # TODO Message: AddFace() takes at most 4 arguments (6 given)
-            if len(self.composition_nodes) == 5:
-                cycle_mesh.Faces.AddFace(0, 1, 2, 3, 4)
-
-            elif len(self.composition_nodes) == 6:
-                cycle_mesh.Faces.AddFace(0, 1, 2, 5)
-                cycle_mesh.Faces.AddFace(2, 3, 4, 5)
-
-            elif len(self.composition_nodes) == 7:
-                cycle_mesh.Faces.AddFace(0, 1, 2, 3, 4, 5, 6)
+        # メッシュの構成点が5つ以上の場合はPolyline
+        else:
+            self.generate_cycle_polyline(layer_name)
+            return
 
         cycle_mesh.Normals.ComputeNormals()
         cycle_mesh.Compact()
@@ -65,8 +59,30 @@ class Cycle:
         self.cycle_mesh_guid = scriptcontext.doc.Objects.AddMesh(cycle_mesh)
         rs.ObjectLayer(self.cycle_mesh_guid, layer)
 
+    def generate_cycle_polyline(self, layer_name):
+
+        # Generating cycle polyline
+        points = [node.point for node in self.composition_nodes]  # p1, p2, p3
+        points.append(points[0])  # p1, p2, p3, p1  <- 最後にスタート点に描画する
+        self.cycle_polyline = Polyline(points)
+
+        # Drawing mesh polyline in doc
+        if layer_name == "r-cycle":
+            layer = rs.AddLayer(self.id, [255, 125, 0], True, False, layer_name)
+        elif layer_name == "v-cycle":
+            layer = rs.AddLayer(self.id, [0, 125, 255], True, False, layer_name)
+        else:
+            layer = None
+
+        self.cycle_polyline_guid = scriptcontext.doc.Objects.AddPolyline(self.cycle_polyline)
+        rs.ObjectLayer(self.cycle_polyline_guid, layer)
+
     def delete_cycle_guid(self):
-        rs.DeleteObject(self.cycle_mesh_guid)
+        if self.cycle_polyline_guid:
+            rs.DeleteObject(self.cycle_polyline_guid)
+
+        if self.cycle_mesh_guid:
+            rs.DeleteObject(self.cycle_mesh_guid)
 
     @staticmethod
     def determine_subset_of_two_cycles(test_cycle_nodes, generated_cycles):
